requirements.txt
ultralytics==8.0.184
clearml[s3]==1.13.0
pydantic<2.0.0
# images
opencv-python-headless>=4.8.0.76
Pillow>=9.3.0

# data
albumentations==1.3.1
label-studio-sdk==0.0.30
label-studio-converter==0.0.55
cvat-sdk==2.6.0
roboflow==1.1.4
gsutil==5.25
requests==2.31.0

# utils
rich==13.5.2
tqdm==4.66.1
pyYAML==6.0.1
python-dotenv==1.0.0

# formats models
coremltools==6.3.0
onnx==1.14.1
onnx-simplifier==0.4.33
onnxruntime==1.15.1
openvino-dev>=2023.1.0
# nvidia-tensorrt
# tensorflow 
# tensorflowjs